src/main/webapp/app/entities/food/detail/food-detail.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { ActivatedRoute } from '@angular/router';
import { of } from 'rxjs';

import { DataUtils } from 'app/core/util/data-util.service';

import { FoodDetailComponent } from './food-detail.component';

describe('Component Tests', () => {
  describe('Food Management Detail Component', () => {
    let comp: FoodDetailComponent;
    let fixture: ComponentFixture<FoodDetailComponent>;
    let dataUtils: DataUtils;

    beforeEach(() => {
      TestBed.configureTestingModule({
        declarations: [FoodDetailComponent],
        providers: [
          {
            provide: ActivatedRoute,
            useValue: { data: of({ food: { id: 123 } }) },
          },
        ],
      })
        .overrideTemplate(FoodDetailComponent, '')
        .compileComponents();
      fixture = TestBed.createComponent(FoodDetailComponent);
      comp = fixture.componentInstance;
      dataUtils = TestBed.inject(DataUtils);
    });

    describe('OnInit', () => {
      it('Should load food on init', () => {
        // WHEN
        comp.ngOnInit();

        // THEN
        expect(comp.food).toEqual(jasmine.objectContaining({ id: 123 }));
      });
    });

    describe('byteSize', () => {
      it('Should call byteSize from DataUtils', () => {
        // GIVEN
        spyOn(dataUtils, 'byteSize');
        const fakeBase64 = 'fake base64';

        // WHEN
        comp.byteSize(fakeBase64);

        // THEN
        expect(dataUtils.byteSize).toBeCalledWith(fakeBase64);
      });
    });

    describe('openFile', () => {
      it('Should call openFile from DataUtils', () => {
        // GIVEN
        spyOn(dataUtils, 'openFile');
        const fakeContentType = 'fake content type';
        const fakeBase64 = 'fake base64';

        // WHEN
        comp.openFile(fakeBase64, fakeContentType);

        // THEN
        expect(dataUtils.openFile).toBeCalledWith(fakeBase64, fakeContentType);
      });
    });
  });
});
